Laid out introduction, edited portfolio
Created an introduction for myself

Began an 'about me' section

Set up Portfolio section to include the games I've worked on, ready to link to dedicated pages.

Commented out a large amount of html code since I'm not planning on using it currently, but may do so in the future. So keeping for now until I fully make up my mind.